usm: http2: tests: Migrated raw payload tests to a new file
The migration is a preliminary step in order to allow running the same tests
both for plain text and tls versions.
